font fix, rendering refactoring, components
