+++ teleoscope.ca/src/components/landing/footer.tsx
@@ -1,0 +1,10 @@
+
+
+export default function Footer() {
+    return (
+        <footer className="flex flex-col items-center justify-center gap-4 bg-gray-50 h-24">
+            <span className="text-2xl font-bold">Teleoscope</span>
+        </footer>
+    );  
+}
+
