+++ teleoscope.ca/src/components/landing/hero.tsx
@@ -1,0 +1,16 @@
+
+
+export default function Hero() {
+    return (
+        <section className="hero h-[60vh] flex-shrink-0 bg-red-50">
+        <div className="flex flex-col items-center justify-center h-full gap-4">
+        <span className="text-4xl font-bold">
+            Empower your research with AI
+            </span>
+            <span className="text-2xl font-medium">
+            AI Assisted Qualitative Analysis
+            </span>
+        </div>
+        </section>
+    );
+    }
