+++ teleoscope.ca/src/components/landing/newsletterSection.tsx
@@ -1,0 +1,77 @@
+
+
+"use client";
+import { Button } from "@/components/ui/button"
+import {
+  Form,
+  FormControl,
+  FormDescription,
+  FormField,
+  FormItem,
+  FormLabel,
+  FormMessage,
+} from "@/components/ui/form"
+import { Input } from "@/components/ui/input"
+
+import { zodResolver } from "@hookform/resolvers/zod"
+import { useForm } from "react-hook-form"
+import { z } from "zod"
+import Link from "next/link";
+
+export default function NewsletterSection() {
+    const formSchema = z.object({
+        username: z.string().min(2).max(50),
+      })
+    
+      const form = useForm<z.infer<typeof formSchema>>({
+        resolver: zodResolver(formSchema),
+        defaultValues: {
+          username: "youremail@example.com",
+        },
+      })
+    
+      function onSubmit(values: z.infer<typeof formSchema>) {
+        console.log(values);
+      }
+    
+    return (
+        <div className="flex items-center justify-center px-8 py-2">
+        <div className="grid grid-cols-3 gap-4 rounded bg-gray-200">
+          <div className="flex items-center justify-center text-xl p-4">
+            Stay up to date with the latest developments in research and join our community with our newsletter.
+          </div>
+          
+          <div className="p-8">
+            <Form {...form}>
+              <form onSubmit={form.handleSubmit(onSubmit)} className="space-y-8">
+                <FormField
+                  control={form.control}
+                  name="username"
+                  render={({ field }) => (
+                    <FormItem>
+                      <FormLabel>Username</FormLabel>
+                      <FormControl>
+                        <Input placeholder="shadcn" {...field} />
+                      </FormControl>
+                      <FormDescription>
+                        This is your public display name.
+                      </FormDescription>
+                      <FormMessage />
+                    </FormItem>
+                  )}
+                />
+                <Button type="submit">Submit</Button>
+              </form>
+            </Form>
+          </div>
+          <div className="flex items-center justify-center p-4">
+            <Button variant="teleoscopeBlue">
+              <Link href="/signup" legacyBehavior passHref>
+                <h2 className="text-lg">Sign Up</h2>
+              </Link>
+            </Button>
+          </div>
+        </div>
+      </div>
+    );
+}
